requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
#-e .